Added discord integration
Added a discord bot and docs to set up a disocrd bot, as well as fixed some issues with the sequelize tables including score. Added tables for discord users and channels as well. However there are some conflicts with the query of retrieving renderable links to display on the main page