Fixes error when selecting `ClusterIP` service type
TL;DR
-----

Correctly sets the Helm value when user chooses `ClusterIP`

Details
-------

Noticed a bug when I deployed with the `ClusterIP` service. I'd
changed the values for the various service type values in the
config to snake case, and that value was being passed on directly
when the selected type was `ClusterIP`. This change fixes that
bug.