Enables building and deploying using airgap features (#81)
TL;DR
-----

Allows Slackernews to build and deploy with airgap support

Details
-------

Updates the Slackernews chart and KOTS configuration so that it can be
deployed using KOTS, EC, and Helm airgap capabilities. It seems odd to talk
about deploying an application depending on access to Slack as an air-gapped
application, but that thinking doesn't take into account the use case that our
airgap support also enables: 

1. Deploying into a network that limits access to outside image registries
2. Security requirements that specify that images must be pulled from an
   internal registry
3. Security policy requiring images to be scanned before deployment, which is
   often accomplished using an internal registry

With this change, Slackernews can be deployed either with EC or KOTS airgap
bundles or using the enterprise portal instructions for Helm airgap.