Uses proxy for all images
TL;DR
-----

Follows new guidance and default to the proxy regsitry

Details
-------

Adopts our new recommendation to use the proxy registry as the default
for all images directly in the Helm chart. It uses the proxy's anonymous
capabilities for the NGINX and Postgres images, while it's application
specific for the Slackernews image.

Also specifies the required image pull secret. This means the chart will
only be usable when pulled/installed from Replicated.