Aligns native libraries (#110)
TL;DR
-----

Fixes bug with SQLite native libraries

Details
-------

Corrects a defect where the SQLite libraries were not loading
correctly in our SecureBuild-based image. The builder layers we used
were Alpine-based, which lead to an incompatibility with the
underlying SQLite library's native code. This led to a runtime
exception and `500` errors all around. The new build-layers are Debian
based and avoid the issue.