Removes SecureBuild image references
TL;DR
-----

Unwinds the change to use SecureBuild images

Details
-------

Responds to image issues with SecureBuild by removing references to
the images we were using. The Node and NGINX images are getting
linking errors with a missing dependency, and the Postgres iamges is
missing the shell script it refers to as the entrypoint. Based on the
symptoms in the other images, I would suspect that the Postgres image
will also have the library issue if the script were present.

To be parepare after the images are fixed, I've left in the build
changes we made to accomodate using SecureBuild for some images.